API_Operaciones/requirements.txt
asgiref==3.8.1
attrs==24.3.0
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.0.10
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-api-logger==1.1.16
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mssql-django==1.5
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
webencodings==0.5.1
gunicorn
